Move expression sorting out of lower loops.
