+++ caffe2/CMakeLists.txt
@@ -548,6 +548,7 @@ if(NOT INTERN_BUILD_MOBILE OR NOT BUILD_CAFFE2_MOBILE)
       ${TORCH_SRC_DIR}/csrc/jit/codegen/cuda/kernel_ir_printer.cpp
       ${TORCH_SRC_DIR}/csrc/jit/codegen/cuda/lower_alias_memory.cpp
       ${TORCH_SRC_DIR}/csrc/jit/codegen/cuda/lower_allocation.cpp
+      ${TORCH_SRC_DIR}/csrc/jit/codegen/cuda/lower_expr_sort.cpp
       ${TORCH_SRC_DIR}/csrc/jit/codegen/cuda/lower_index.cpp
       ${TORCH_SRC_DIR}/csrc/jit/codegen/cuda/lower_insert_syncs.cpp
       ${TORCH_SRC_DIR}/csrc/jit/codegen/cuda/lower_loops.cpp
